Package structure renaming.
