+++ shared/src/commonMain/kotlin/template/shared/data/local/DatabaseDriverFactory.kt
@@ -1,4 +1,4 @@
-package template.shared
+package template.shared.data.local
 
 import app.cash.sqldelight.db.SqlDriver
