+++ shared/src/commonMain/kotlin/template/shared/data/remote/BaseKtorClient.kt
@@ -1,4 +1,4 @@
-package template.shared
+package template.shared.data.remote
 
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
