+++ shared/src/commonMain/kotlin/template/shared/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package template.shared.theme
+package template.shared.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
