+++ shared/src/commonMain/kotlin/template/shared/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package template.shared.theme
+package template.shared.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
