+++ shared/src/iosMain/kotlin/template/shared/data/local/DatabaseDriverFactory.ios.kt
@@ -1,9 +1,10 @@
 @file:Suppress("ktlint:standard:filename")
 
-package template.shared
+package template.shared.data.local
 
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.native.NativeSqliteDriver
+import template.shared.data.local.DatabaseDriverFactory
 
 /**
  * Create an instance of a [SqlDriver] for the iOS platform.
